+++ app/services/weeks/coming_service.rb
@@ -6,10 +6,12 @@ module Weeks
 
     def initialize(
       lineup_create_service: Lineups::CreateService,
-      games_players_create_service: FantasySports::Container['services.games.players.create_for_game']
+      games_players_create_service: FantasySports::Container['services.games.players.create_for_game'],
+      update_game_difficulties_job: Games::DifficultyUpdateJob
     )
       @lineup_create_service = lineup_create_service
       @games_players_create_service = games_players_create_service
+      @update_game_difficulties_job = update_game_difficulties_job
     end
 
     def call(week:)
@@ -20,6 +22,7 @@ module Weeks
       week.update!(status: Week::COMING)
       create_games_players(week)
       create_lineups(week)
+      update_game_difficulties(week)
     end
 
     private
@@ -32,6 +35,10 @@ module Weeks
       week.season.fantasy_teams.completed.find_each { |fantasy_team|
         @lineup_create_service.call(fantasy_team: fantasy_team, week: week)
       }
+    end
+
+    def update_game_difficulties(week)
+      @update_game_difficulties_job.perform_later(week_id: week.id)
     end
   end
 end